fix(platform): schematics
