+++ libs/platform/schematics/ng-add/schema.ts
@@ -1,0 +1,4 @@
+export interface Schema {
+    project: string;
+    translations: boolean;
+}
